INDTR 20912 5.7.2. App.js - PDP modifications
Created Confirmation controller that displays the wishlist Modal to ask for confirmation if user is not logged in.
Modified index.js to prevent event default if button Add to wishlist clicked. If user is not logged in it will use Dialog.open and call the COnfirmation controller to display amodal and ask for confirmation. If user logged in it will proceed to add the item to wishlist